Add files via upload
A very crude model of a kerbal without helmet, with the origin corresponding to the location of an IVA InternalSeat transform. In other words, if you put this kerbonaut in your IVA model temporarily, and line it up to where you want your crew member to be, then a transform with the same location and orientation will put them there. Remember that for this model and the InternalSeat transform itself, z faces forward, and Y faces upward